+++ Assets/Shaders/Resources/IndirectDepth.shader
@@ -31,7 +31,7 @@
                 return o;
             }
 
-            float frag (v2f i) : SV_Target
+            half frag (v2f i) : SV_Target
             {
                 return Linear01Depth(i.vertex.z);
             }
